// src/types/certificateData.ts
export type certificateData = {
    numeroPoliza: string;
    fechaEmision: string;
    horaEmision: string;
    fechaInicioCovertura: string;
    fechaFinCovertura: string;
    vehiculoPlaca: string;
    nombres: string;
    vehiculoClaseDescripcion: string;
    vehiculoModeloNombre: string;
    vehiculoMarcaNombre: string;
    precio: string;
    vehiculoUsoDescripcion: string;
    vehiculoVin: string;
}